Handles larger packets better. Still got odd disconnection bug.

+++ WebSockets/WebSocketMessage.cs
@@ -17,7 +17,7 @@ namespace WebSockets
         public bool Reserved3;
         public byte Opcode;
         public bool Mask;
-        public int PayloadLength;
+        public long PayloadLength;
         public byte[] MaskingKey;
         public List<byte> Data;
